Add SQL query to get all devices through sql connector

+++ mysql/sql_scripts/startup.sql
@@ -2,7 +2,7 @@ CREATE TABLE Devices (
 	DeviceID      INT NOT NULL AUTO_INCREMENT,
 	Name          VARCHAR(255) NOT NULL,
 	IP            VARCHAR(255) NOT NULL,
-	GeraeteNummer  DECIMAL NOT NULL,
+	GeraeteNummer  INTEGER NOT NULL,
 	Settings      VARCHAR(255) NOT NULL,
 	PRIMARY KEY (DeviceID)
 );
